RDKB-47543 radio channel, channelwidth, dfs validation in hal
Reason for change: added validation check to compare the channel, channelwidth and dfs, if the input is having invalid channel for
corresponding channelwidth/dfs the apply should fail and report an error in wifiWebConfig log.
Test Procedure: using dmcli and ovsh commands, change the channel and if the input channel is invalid apply should fail.
Risks: Low
Priority: P1

Signed-off-by: Ramakanth Taraka <ramakanth.taraka@sky.uk>
Change-Id: I4abd9ab9298e6c34716b7fad154bb1c2141eb72c